+++ beacon-chain/state/stategen/replay.go
@@ -235,79 +235,6 @@ func ReplayProcessSlots(ctx context.Context, state state.BeaconState, slot types
 	return state, nil
 }
 
-// This finds the last saved block in DB from searching backwards from input slot,
-// it returns the block root and the slot of the block.
-// This is used by both hot and cold state management.
-func (s *State) lastSavedBlock(ctx context.Context, slot types.Slot) ([32]byte, types.Slot, error) {
-	ctx, span := trace.StartSpan(ctx, "stateGen.lastSavedBlock")
-	defer span.End()
-
-	// Handle the genesis case where the input slot is 0.
-	if slot == 0 {
-		gRoot, err := s.genesisRoot(ctx)
-		if err != nil {
-			return [32]byte{}, 0, err
-		}
-		return gRoot, 0, nil
-	}
-
-	lastSaved, err := s.beaconDB.HighestSlotBlocksBelow(ctx, slot)
-	if err != nil {
-		return [32]byte{}, 0, err
-	}
-
-	// Given this is used to query canonical block. There should only be one saved canonical block of a given slot.
-	if len(lastSaved) != 1 {
-		return [32]byte{}, 0, fmt.Errorf("highest saved block does not equal to 1, it equals to %d", len(lastSaved))
-	}
-	if lastSaved[0] == nil || lastSaved[0].IsNil() || lastSaved[0].Block().IsNil() {
-		return [32]byte{}, 0, nil
-	}
-	r, err := lastSaved[0].Block().HashTreeRoot()
-	if err != nil {
-		return [32]byte{}, 0, err
-	}
-
-	return r, lastSaved[0].Block().Slot(), nil
-}
-
-// This finds the last saved state in DB from searching backwards from input slot,
-// it returns the block root of the block which was used to produce the state.
-// This is used by both hot and cold state management.
-func (s *State) lastSavedState(ctx context.Context, slot types.Slot) (state.ReadOnlyBeaconState, error) {
-	ctx, span := trace.StartSpan(ctx, "stateGen.lastSavedState")
-	defer span.End()
-
-	// Handle the genesis case where the input slot is 0.
-	if slot == 0 {
-		return s.beaconDB.GenesisState(ctx)
-	}
-
-	lastSaved, err := s.beaconDB.HighestSlotStatesBelow(ctx, slot+1)
-	if err != nil {
-		return nil, err
-	}
-
-	// Given this is used to query canonical state. There should only be one saved canonical block of a given slot.
-	if len(lastSaved) != 1 {
-		return nil, fmt.Errorf("highest saved state does not equal to 1, it equals to %d", len(lastSaved))
-	}
-	if lastSaved[0] == nil {
-		return nil, errUnknownState
-	}
-
-	return lastSaved[0], nil
-}
-
-// This returns the genesis root.
-func (s *State) genesisRoot(ctx context.Context) ([32]byte, error) {
-	b, err := s.beaconDB.GenesisBlock(ctx)
-	if err != nil {
-		return [32]byte{}, err
-	}
-	return b.Block().HashTreeRoot()
-}
-
 // Given the start slot and the end slot, this returns the finalized beacon blocks in between.
 // Since hot states don't have finalized blocks, this should ONLY be used for replaying cold state.
 func (s *State) loadFinalizedBlocks(ctx context.Context, startSlot, endSlot types.Slot) ([]block.SignedBeaconBlock, error) {
