+++ beacon-chain/state/stategen/replay_test.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"testing"
 
-	types "github.com/prysmaticlabs/eth2-types"
 	"github.com/prysmaticlabs/prysm/beacon-chain/core/blocks"
 	"github.com/prysmaticlabs/prysm/beacon-chain/db"
 	testDB "github.com/prysmaticlabs/prysm/beacon-chain/db/testing"
@@ -342,135 +341,6 @@ func TestLoadBlocks_BadStart(t *testing.T) {
 	require.NoError(t, err)
 	_, err = s.LoadBlocks(ctx, 0, 5, roots[8])
 	assert.ErrorContains(t, "end block roots don't match", err)
-}
-
-func TestLastSavedBlock_Genesis(t *testing.T) {
-	beaconDB := testDB.SetupDB(t)
-	ctx := context.Background()
-	s := &State{
-		beaconDB:      beaconDB,
-		finalizedInfo: &finalizedInfo{slot: 128},
-	}
-
-	gBlk := util.NewBeaconBlock()
-	gRoot, err := gBlk.Block.HashTreeRoot()
-	require.NoError(t, err)
-	require.NoError(t, s.beaconDB.SaveBlock(ctx, wrapper.WrappedPhase0SignedBeaconBlock(gBlk)))
-	require.NoError(t, s.beaconDB.SaveGenesisBlockRoot(ctx, gRoot))
-
-	savedRoot, savedSlot, err := s.lastSavedBlock(ctx, 0)
-	require.NoError(t, err)
-	assert.Equal(t, types.Slot(0), savedSlot, "Did not save genesis slot")
-	assert.Equal(t, savedRoot, savedRoot, "Did not save genesis root")
-}
-
-func TestLastSavedBlock_CanGet(t *testing.T) {
-	beaconDB := testDB.SetupDB(t)
-	ctx := context.Background()
-	s := &State{
-		beaconDB:      beaconDB,
-		finalizedInfo: &finalizedInfo{slot: 128},
-	}
-
-	b1 := util.NewBeaconBlock()
-	b1.Block.Slot = s.finalizedInfo.slot + 5
-	require.NoError(t, s.beaconDB.SaveBlock(ctx, wrapper.WrappedPhase0SignedBeaconBlock(b1)))
-	b2 := util.NewBeaconBlock()
-	b2.Block.Slot = s.finalizedInfo.slot + 10
-	require.NoError(t, s.beaconDB.SaveBlock(ctx, wrapper.WrappedPhase0SignedBeaconBlock(b2)))
-	b3 := util.NewBeaconBlock()
-	b3.Block.Slot = s.finalizedInfo.slot + 20
-	require.NoError(t, s.beaconDB.SaveBlock(ctx, wrapper.WrappedPhase0SignedBeaconBlock(b3)))
-
-	savedRoot, savedSlot, err := s.lastSavedBlock(ctx, s.finalizedInfo.slot+100)
-	require.NoError(t, err)
-	assert.Equal(t, s.finalizedInfo.slot+20, savedSlot)
-	wantedRoot, err := b3.Block.HashTreeRoot()
-	require.NoError(t, err)
-	assert.Equal(t, wantedRoot, savedRoot, "Did not save correct root")
-}
-
-func TestLastSavedBlock_NoSavedBlock(t *testing.T) {
-	beaconDB := testDB.SetupDB(t)
-	ctx := context.Background()
-	s := &State{
-		beaconDB:      beaconDB,
-		finalizedInfo: &finalizedInfo{slot: 128},
-	}
-
-	root, slot, err := s.lastSavedBlock(ctx, s.finalizedInfo.slot+1)
-	require.NoError(t, err)
-	if slot != 0 && root != [32]byte{} {
-		t.Error("Did not get wanted block")
-	}
-}
-
-func TestLastSavedState_Genesis(t *testing.T) {
-	beaconDB := testDB.SetupDB(t)
-	ctx := context.Background()
-	s := &State{
-		beaconDB:      beaconDB,
-		finalizedInfo: &finalizedInfo{slot: 128},
-	}
-
-	gBlk := util.NewBeaconBlock()
-	gState, err := util.NewBeaconState()
-	require.NoError(t, err)
-	gRoot, err := gBlk.Block.HashTreeRoot()
-	require.NoError(t, err)
-	require.NoError(t, s.beaconDB.SaveBlock(ctx, wrapper.WrappedPhase0SignedBeaconBlock(gBlk)))
-	require.NoError(t, s.beaconDB.SaveGenesisBlockRoot(ctx, gRoot))
-	require.NoError(t, s.beaconDB.SaveState(ctx, gState, gRoot))
-
-	savedState, err := s.lastSavedState(ctx, 0)
-	require.NoError(t, err)
-	require.DeepSSZEqual(t, gState.InnerStateUnsafe(), savedState.InnerStateUnsafe())
-}
-
-func TestLastSavedState_CanGet(t *testing.T) {
-	beaconDB := testDB.SetupDB(t)
-	ctx := context.Background()
-	s := &State{
-		beaconDB:      beaconDB,
-		finalizedInfo: &finalizedInfo{slot: 128},
-	}
-
-	b1 := util.NewBeaconBlock()
-	b1.Block.Slot = s.finalizedInfo.slot + 5
-	require.NoError(t, s.beaconDB.SaveBlock(ctx, wrapper.WrappedPhase0SignedBeaconBlock(b1)))
-	b2 := util.NewBeaconBlock()
-	b2.Block.Slot = s.finalizedInfo.slot + 10
-	require.NoError(t, s.beaconDB.SaveBlock(ctx, wrapper.WrappedPhase0SignedBeaconBlock(b2)))
-	b2Root, err := b2.Block.HashTreeRoot()
-	require.NoError(t, err)
-	st, err := util.NewBeaconState()
-	require.NoError(t, err)
-	require.NoError(t, st.SetSlot(s.finalizedInfo.slot+10))
-
-	require.NoError(t, s.beaconDB.SaveState(ctx, st, b2Root))
-	b3 := util.NewBeaconBlock()
-	b3.Block.Slot = s.finalizedInfo.slot + 20
-	require.NoError(t, s.beaconDB.SaveBlock(ctx, wrapper.WrappedPhase0SignedBeaconBlock(b3)))
-
-	savedState, err := s.lastSavedState(ctx, s.finalizedInfo.slot+100)
-	require.NoError(t, err)
-	require.DeepSSZEqual(t, st.InnerStateUnsafe(), savedState.InnerStateUnsafe())
-}
-
-func TestLastSavedState_NoSavedBlockState(t *testing.T) {
-	beaconDB := testDB.SetupDB(t)
-	ctx := context.Background()
-	s := &State{
-		beaconDB:      beaconDB,
-		finalizedInfo: &finalizedInfo{slot: 128},
-	}
-
-	b1 := util.NewBeaconBlock()
-	b1.Block.Slot = 127
-	require.NoError(t, s.beaconDB.SaveBlock(ctx, wrapper.WrappedPhase0SignedBeaconBlock(b1)))
-
-	_, err := s.lastSavedState(ctx, s.finalizedInfo.slot+1)
-	assert.ErrorContains(t, errUnknownState.Error(), err)
 }
 
 // tree1 constructs the following tree:
