+++ proto/prysm/v1alpha1/beacon_chain.proto
@@ -150,16 +150,6 @@ service BeaconChain {
         };
     }
 
-    // Retrieve information about the weak subjectivity of the beacon chain from the view of
-    // the beacon chain node.
-    //
-    // This includes the weak subjectivity block root, state root and epoch number.
-    rpc GetWeakSubjectivityCheckpoint(google.protobuf.Empty) returns (WeakSubjectivityCheckpoint) {
-        option (google.api.http) = {
-            get: "/eth/v1alpha1/beacon/weak_subjectivity_checkpoint"
-        };
-    }
-
     // Retrieve the beacon chain committees for a given epoch.
     //
     // If no filter criteria is specified, the response returns
@@ -907,13 +897,4 @@ message IndividualVotesRespond {
     }
 
     repeated IndividualVote individual_votes = 1;
-}
-
-message WeakSubjectivityCheckpoint {
-    // The block root of weak subjectivity checkpoint.
-    bytes block_root = 1;
-    // The state root of weak subjectivity checkpoint.
-    bytes state_root = 2;
-    // The epoch of weak subjectivity checkpoint.
-    uint64 epoch = 3 [(ethereum.eth.ext.cast_type) = "github.com/prysmaticlabs/eth2-types.Epoch"];
 }
